Commit again!

+++ Form1.cs
@@ -303,7 +303,7 @@ namespace DietSentry
                         });
                         context.SaveChanges();
 
-                        labelInfo.Text = "Added new food item " + addedFoodItem.FoodDescription;
+                        labelInfoFood.Text = "Added new food item " + addedFoodItem.FoodDescription;
                     }
                     else if (foodType == 2)
                     {
@@ -321,7 +321,7 @@ namespace DietSentry
                 }
                 else // addition of food item cancelled
                 {
-                    labelInfo.Text = "Addition of food item cancelled";
+                    labelInfoFood.Text = "Addition of food item cancelled";
                 }
             }
             else if (e.KeyCode == Keys.F2) // Editing selected food item
@@ -423,11 +423,11 @@ namespace DietSentry
                         {
                             ActOnFoodFilteringStates(1);
                         }
-                        labelInfo.Text = "Edited food item " + editedFoodId;
+                        labelInfoFood.Text = "Edited food item " + editedFoodId;
                     }
                     else // editing of food item cancelled
                     {
-                        labelInfo.Text = "Editing of food item " + editedFoodId + " cancelled";
+                        labelInfoFood.Text = "Editing of food item " + editedFoodId + " cancelled";
                     }
                 }
             }
@@ -474,7 +474,7 @@ namespace DietSentry
                         else // proceed to create copy of food item in Solid form
                         {
                             // Creating new FoodDescription, with the same trucated base but extension varying depending on original food type
-                            string sDescriptionDensity = editedFoodDescription + " {densitye=" + string.Format("{0:N3}", densityOfFood) + "g/ml}";
+                            string sDescriptionDensity = editedFoodDescription + " {density=" + string.Format("{0:N3}", densityOfFood) + "g/ml}";
                             if (fT == 1) // dealing with "public" liquid
                             {
                                 editedFoodDescription = sDescriptionDensity;
@@ -545,21 +545,33 @@ namespace DietSentry
             // can delete row only if the totals filter check boxed is unchecked
             if ((!checkBoxDailyTotals.Checked))
             {
-                // determine row to be deleted from Eaten table by accessing it in its dataGrid
-                var foodItem = (Eaten)this.dataGridViewEaten.CurrentRow.DataBoundItem;
+                // Initializes the variables to pass to the MessageBox.Show method.
+                string message = "Are you sure you wish to delete the selected eaten food?";
+                string caption = "DELETE CONFIRMATION";
+                DialogResult result = MessageBox.Show(message, caption, MessageBoxButtons.YesNo, MessageBoxIcon.None, MessageBoxDefaultButton.Button1);
 
-                if (foodItem != null)
+                if (result == DialogResult.Yes)
                 {
-                    using var context = new FoodsContext();
-                    // gain direct access to selected entry in Eaten table
-                    var FoodSelected = context.Eaten.Single(b => b.EatenId == foodItem.EatenId);
+                    // determine row to be deleted from Eaten table by accessing it in its dataGrid
+                    var foodItem = (Eaten)this.dataGridViewEaten.CurrentRow.DataBoundItem;
 
-                    // delete this row from Eaten table
-                    context.Eaten.Remove(FoodSelected);
-                    context.SaveChanges();
+                    if (foodItem != null)
+                    {
+                        using var context = new FoodsContext();
+                        // gain direct access to selected entry in Eaten table
+                        var FoodSelected = context.Eaten.Single(b => b.EatenId == foodItem.EatenId);
 
-                    // refresh Eaten data grid view while maintaining filter status
-                    ActOnEatenFoodFilteringStates();
+                        // delete this row from Eaten table
+                        context.Eaten.Remove(FoodSelected);
+                        context.SaveChanges();
+
+                        // refresh Eaten data grid view while maintaining filter status
+                        ActOnEatenFoodFilteringStates();
+                    }
+                }
+                else
+                {
+                    labelInfoEaten.Text = "Deletion of selected eaten food cancelled!";
                 }
             }
             else // prevent row deletion and inform user of this by displaying Message Box
@@ -849,7 +861,7 @@ namespace DietSentry
                     }
                     else
 
-                        labelInfo.Text = "Deleted food item " + FoodSelectedId;
+                        labelInfoFood.Text = "Deleted food item " + FoodSelectedId;
 
                     // refresh Foods data grid view while maintaining filter status - it is done in this complicated way
                     // because it crashes under some patricular circumstances !
@@ -862,7 +874,7 @@ namespace DietSentry
             }
             else
             {
-                labelInfo.Text = "Deletion of selected food item cancelled";
+                labelInfoFood.Text = "Deletion of selected food item cancelled";
             }
         }
